+++ models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from pathlib import Path
 
 
-VERSION = "2.3.4"
+VERSION = "2.3.5"
 
 
 @dataclass(frozen=True)